feat(tokens): refactor tokens group and split into separated spec files
