+++ tests/tokens/token-group/token-group.spec.ts-snapshots/linux/chrome/token-group-primary-nested.png
