+++ tests/tokens/token-group/token-group.spec.ts-snapshots/linux/firefox/token-group-primary-nested.png
